Renamed files and some code refactoring
